+++ src/test/java/com/tabisketch/api/CreatePlaceAPIControllerTest.java
@@ -1,10 +1,9 @@
-package com.tabisketch.controller;
+package com.tabisketch.api;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.tabisketch.bean.form.CreateDayForm;
-import com.tabisketch.bean.form.ExampleCreateDayForm;
-import com.tabisketch.bean.response.CreateDayResponse;
-import com.tabisketch.service.implement.CreateDayService;
+import com.tabisketch.bean.form.ExampleCreatePlaceForm;
+import com.tabisketch.bean.response.CreatePlaceResponse;
+import com.tabisketch.service.ICreatePlaceService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -18,27 +17,27 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
-@WebMvcTest(CreateDayAPIController.class)
-public class CreateDayAPIControllerTest {
+@WebMvcTest(CreatePlaceAPIController.class)
+public class CreatePlaceAPIControllerTest {
     @Autowired
     private MockMvc mockMvc;
     @Autowired
     private ObjectMapper objectMapper;
     @MockitoBean
-    private CreateDayService createDayService;
+    private ICreatePlaceService createPlaceService;
 
     @Test
     @WithMockUser
     public void testPost() throws Exception {
-        final int dayId = 1;
-        when(this.createDayService.execute(any())).thenReturn(dayId);
+        final int placeId = 1;
+        when(this.createPlaceService.execute(any())).thenReturn(placeId);
 
-        final var createDayForm = ExampleCreateDayForm.generate();
-        final String responseJson = this.objectMapper.writeValueAsString(CreateDayResponse.success(dayId));
+        final var createPlaceForm = ExampleCreatePlaceForm.generate();
+        final String responseJson = this.objectMapper.writeValueAsString(CreatePlaceResponse.success(placeId));
 
         this.mockMvc.perform(MockMvcRequestBuilders
-                        .post("/api/create-day")
-                        .flashAttr("createDayForm", createDayForm)
+                        .post("/api/create-place")
+                        .flashAttr("createPlaceForm", createPlaceForm)
                         .with(SecurityMockMvcRequestPostProcessors.csrf())
                 ).andExpect(MockMvcResultMatchers.status().isOk())
                 .andExpect(MockMvcResultMatchers.content().json(responseJson));
